Created a process-based supervisor on windows
Our home-made Windows supervisor is know using processes launched using
`subprocess.Popen` instead of `multiprocessing`. It gets closer to the
behaviour of `supervisord` on Linux and leads to a much smaller code for
the supervisor itself. On the other hand it makes it a bit harder to
have a windows-specific behaviour for our subprocesses (like JMXFetch)
but cross-compatibility issues should rather be taken care of in the
code of these subprocesses themselves instead of doing it in the upper
layer.